refactor(aia): move Iid and MajorIid structs into independent module `iid`
