+++ library/riscv-aia/src/register/vstopi.rs
@@ -1,5 +1,7 @@
 //! Virtual supervisor top interrupt (vstopi)
 
+use crate::iid::MajorIid;
+
 riscv::read_only_csr! {
     /// Virtual supervisor top interrupt register.
     Vstopi: 0xEB0,
@@ -8,9 +10,12 @@ riscv::read_only_csr! {
 
 impl Vstopi {
     #[inline]
-    pub const fn iid(self) -> Option<crate::Iid> {
-        let bits = (self.bits & 0x0FFF_0000) >> 16;
-        crate::Iid::new(bits as u16)
+    pub const fn iid(self) -> Option<MajorIid> {
+        let number = ((self.bits & 0x0FFF_0000) >> 16) as u16;
+        match number {
+            0 => None,
+            _ => Some(MajorIid::new(number)),
+        }
     }
 
     #[inline]
@@ -25,7 +30,7 @@ mod tests {
 
     #[test]
     fn vstopi_parsing() {
-        let iid_num: u16 = 0x456;
+        let iid_num: u16 = 0xFFF;
         let iprio: u8 = 0xFF;
         let bits: usize = ((iid_num as usize) << 16) | (iprio as usize);
         let reg = Vstopi::from_bits(bits);
